Separate gpact and sfc event watchers
